backend/requirements.txt
attrs==21.2.0
certifi==2021.5.30
chardet==3.0.4
click==8.0.1
dnspython==2.1.0
Flask==2.0.1
Flask-Cors==3.0.10
Flask-PyMongo==2.3.0
idna==2.5
iniconfig==1.1.1
itsdangerous==2.0.1
Jinja2==3.0.1
logger==1.4
MarkupSafe==2.0.1
mongoengine==0.23.1
mongomock==3.23.0
packaging==21.0
pluggy==1.0.0
py==1.10.0
pymongo==3.12.0
pyparsing==2.4.7
pytest==6.2.5
python-dotenv==0.19.0
PyYAML==5.4.1
requests==2.18.1
sentinels==1.0.0
six==1.16.0
toml==0.10.2
urllib3==1.21.1
Werkzeug==2.0.1
bcrypt==3.2.0
boto=2.49.0
waitress==2.0.0
gunicorn
